Creado Objet Mapper y cambios en clases
Se ha creado los Object Mapper para las clases Temperatura y Pulsaciones, así como diferentes cambios realizados.
